refactor: 리듀서 분리
