+++ prepare/front/components/UserProfile.js
@@ -1,7 +1,7 @@
 import React, { useCallback } from "react";
 import { Avatar, Card, Button } from "antd";
 import { useDispatch } from "react-redux";
-import { logoutAction } from "../reducers";
+import { logoutAction } from "../reducers/user";
 
 const dummy = {
   nickname: "지니",
